requirements.txt
absl-py==0.9.0
anykeystore==0.2
apex==0.1
astor==0.8.1
attrs==19.3.0
bdlb==0.0.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.2.2
cryptacular==1.5.5
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
dominate==2.4.0
future==0.18.2
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.51.0
grpcio==1.26.0
h5py==2.10.0
hupper==1.9.1
idna==2.8
imageio==2.6.1
joblib==0.14.1
kaggle==1.5.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
mock==3.0.5
networkx==2.4
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
PasteDeploy==2.0.1
pbkdf2==1.3
Pillow==6.2.1
plaster==1.0
plaster-pastedeploy==0.7
promise==2.3
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.5
pyramid==1.10.4
pyramid-mailer==0.15.1
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
PyWavelets==1.1.1
repoze.sendmail==4.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.2.0
seaborn==0.10.0
six==1.13.0
SQLAlchemy==1.3.11
tensorboard==1.13.1
tensorflow-datasets==2.0.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-metadata==0.21.0
termcolor==1.1.0
text-unidecode==1.3
tfp-nightly==0.10.0.dev20200127
torch==1.0.1
torchvision==0.2.2
tqdm==4.42.0
transaction==2.4.0
translationstring==1.3
urllib3==1.24.3
velruse==1.1.1
venusian==3.0.0
WebOb==1.8.5
Werkzeug==0.16.1
wrapt==1.11.2
WTForms==2.2.1
wtforms-recaptcha==0.3.2
zope.deprecation==4.4.0
zope.interface==4.7.1
zope.sqlalchemy==1.2
